Update Automate-Module.psm1
Added check to restart service if last check-in is showing more than 10 minutes ago and service is running. 
This is to correct a situations where script was uninstalling and reinstalling the agent service on some machines during boot even though agent was running but agent had not successfully checked in yet after boot. 
This was causing duplicate agents on some of our machines.